users system, #5

+++ inc/models/meal.php
@@ -2,14 +2,14 @@
 
 namespace Diab;
 
-class Meal {
+class Meal extends Model {
   public $id;
   public $date;
   public $date_end;
   public $active;
   public $servings = null;
 
-  private static function _initMeal($row, $getServings = false) {
+  private function _initMeal($row, $getServings = false) {
     $meal = new Meal();
     foreach($row as $k => $v) {
       if($k === "date" || $k === "date_end") {
@@ -18,13 +18,14 @@ class Meal {
       $meal->{$k} = $v;
     }
     if($getServings) {
-      $meal->servings = Serving::getForMeal($meal->id);
+      $serving = new Serving($this->userId);
+      $meal->servings = $serving->getForMeal($meal->id);
     }
     return $meal;
   }
 
-  public static function getActive() {
-    $rows = DB::inst()->to_array("select * from meals where `active` = 1");
+  public function getActive() {
+    $rows = DB::inst()->to_array("select * from meals where `active` = 1 AND user_id = #d", [$this->userId]);
     if(!$rows) {
       return null;
     }
@@ -36,8 +37,8 @@ class Meal {
     return $meal;
   }
 
-  public static function get($id) {
-    $rows = DB::inst()->to_array("select * from meals where `id` = #d", [$id]);
+  public function get($id) {
+    $rows = DB::inst()->to_array("select * from meals where `id` = #d AND user_id = #d", [$id, $this->userId]);
     if(!$rows) {
       return null;
     }
@@ -49,29 +50,30 @@ class Meal {
     return $meal;
   }
 
-  public static function getAll($params) {
+  public function getAll($params) {
     $query = "
       SELECT
         *
       FROM
         `meals`
       WHERE
-        `active` = 0
+        `active` = 0 AND user_id = #d
       ORDER BY
         `id` DESC
       LIMIT #d, #d
     ";
-    $rows = DB::inst()->to_array($query, [$params["offset"], $params["limit"]]);
+    $rows = DB::inst()->to_array($query, [$this->userId, $params["offset"], $params["limit"]]);
     $meals = [];
     foreach($rows as $row) {
-      $meal = self::_initMeal($row, true);
+      $meal = $this->_initMeal($row, true);
       $meals[] = $meal;
     }
     return $meals;
   }
 
-  public static function end($id) {
-    $servings = Serving::getForMeal($id);
+  public function end($id) {
+    $servingModel = new Serving($this->userId);
+    $servings = $servingModel->getForMeal($id);
     if($servings) {
       $eatDates = array_map(function($s) {
         return $s->eat_date;
@@ -83,32 +85,37 @@ class Meal {
         $date = "'".date("Y-m-d H:i:s", round(min($eatDates)/1000))."'";
         $date_end = "'".date("Y-m-d H:i:s", round(max($eatDates)/1000))."'";
       }
-      DB::inst()->q("update meals set `active` = 0, `date_end` = $date_end, `date` = $date where `id` = #d", [$id]);
+      DB::inst()->q(
+        "update meals set `active` = 0, `date_end` = $date_end, `date` = $date where `id` = #d and user_id = #d",
+        [$id, $this->userId]
+      );
       $res = self::get($id);
     }
     else {
       $res = null;
-      DB::inst()->q("delete from `meals` where `id` = #d", [$id]);
+      DB::inst()->q("delete from `meals` where `id` = #d and user_id = #d", [$id, $this->userId]);
     }
     return $res;
   }
 
-  public static function setCoef($id, $coef) {
-    DB::inst()->q("update meals set `coef` = #s where `id` = #d", [$coef, $id]);
+  public function setCoef($id, $coef) {
+    DB::inst()->q("update meals set `coef` = #s where `id` = #d and user_id = #d", [$coef, $id, $this->userId]);
   }
 
-  public static function add($data) {
-    DB::inst()->q(
-      "INSERT INTO `meals`
-      SET `date` = NOW()"
-    );
+  public function add($data) {
+    DB::inst()->q("
+      INSERT INTO
+        `meals`
+      SET
+        `date` = NOW(), user_id = #d
+    ", [$this->userId]);
     $id = DB::inst()->id();
     if(!$id) {
       throw new \Exception("Fail to create meal");
     }
     if(isset($data["coef"])) {
-      self::setCoef($id, $data["coef"]);
+      $this->setCoef($id, $data["coef"]);
     }
-    return self::get($id);
+    return $this->get($id);
   }
 }
